requirements.txt
google-cloud-bigquery==2.25.1
pandas==1.2.4
streamlit==0.87.0
plotly==5.2.2

